refactor: move customer logic into module of its own
